requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
pydantic==2.12.0
sqlalchemy==2.0.43
alembic==1.16.5
scikit-learn==1.7.2
pandas==2.3.3
joblib==1.5.2
streamlit==1.50.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
apache-airflow==3.1.0
numpy==2.33
requests==2.32.5